refactor: separate clipboard logics into actions
