+++ packages/core/src/web/app/svgedit/operations/clipboard/actions/index.ts
@@ -1,0 +1,5 @@
+export * from './array';
+export * from './clone';
+export * from './copy';
+export * from './cut';
+export * from './paste';
